1.6.2

+++ Minecraft/src/net/minecraft/src/ChunkCoordinates.java
@@ -46,9 +46,9 @@ public class ChunkCoordinates implements Comparable
 	
 	public float getDistanceSquared(int par1, int par2, int par3)
 	{
-		int var4 = posX - par1;
-		int var5 = posY - par2;
-		int var6 = posZ - par3;
+		float var4 = posX - par1;
+		float var5 = posY - par2;
+		float var6 = posZ - par3;
 		return var4 * var4 + var5 * var5 + var6 * var6;
 	}
